Day 05 : Factor 1000 improvement
I was convinced the looping approach was inefficient but wasn't sure how
to make it better because of the need to always "look back" once we find
a pair to delete.

I was reminded of the possibility to use an accumulator stack to
properly pop the characters just like a matching parens problem. The
improvement is pretty steep. I let the part a implementation untouched
to see the difference in time it yields (and part c is part a
"optimized")

Part b uses the quick reaction now, so we don't wait too long