Day09 : Slow in Rust
This implementation of the circle buffer with a VecDeque is decent
enough for part a. But insert() operation is O(n) in a VecDeque and
this is terrible for part b with a much bigger queue.

I suspect it would work, but having a LinkedList is probably better
here, since we have constant time insertion (we have max 7 moves from
the current element in the linked list, if it's circular)