Day 09 : Fast in C++
Rust makes it *very* hard to have self-referencing structs. It makes it
very difficult, even with intrusive_collections crate, to actually have
a mutable cursor on a mutable linked list to make the implementation of
part b fast.

I made it in C++ to show what I want to be able to do more easily in
Rust. (Also, part b took 1h in release in Rust, so you should really go
for the C++ implementation here ; the data structure is much better with
the constant time insertion around a maintained iterator)