Day 14 : C++ with one caveat
I cannot ensure Part B works with patterns starting with 0, given that I
do all comparison/insertion operations using numbers, the leading 0 is
eaten away.

No need to fix this until I really need it (and I do not, since my
puzzle input has no leading 0), so this solution is working well.

I recycled the logic from Day 9 (circular list with mutable cursors on
it), with adding just another cursor for the new player, and this time I
am actually using the `begin()` and `rbegin()` methods of the underlying
`std::list<>` implementation, since we are interested in absolute
positions when it comes to compute relevant parts of the game state